claude_bank/app/agents/payment-agent-v3-a2a/pyproject.toml
[project]
name = "payment-agent-v3-a2a"
version = "3.0.0"
description = "Payment Agent V3 A2A Microservice - 2-tool transfer flow (prepareTransfer + executeTransfer)"
requires-python = ">=3.11"

dependencies = [
    # Agent Framework
    "agent-framework-core==1.0.0b251223",
    "agent-framework-azure-ai==1.0.0b251223",
    "agent-framework-a2a==1.0.0b251223",

    # Azure AI Foundry V2 SDK
    "azure-ai-projects>=2.0.0b2",
    "azure-ai-agents==1.2.0b5",

    # A2A SDK
    "a2a-sdk>=0.3.5",

    # Azure Identity
    "azure-identity>=1,<2",

    # Pydantic
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",

    # OpenAI SDK
    "openai>=1.99.0",

    # MCP support
    "mcp[ws]>=1.23",

    # FastAPI for A2A HTTP server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # HTTP client
    "aiohttp",
    "httpx>=0.27.0",

    # Environment variables
    "python-dotenv>=1.0.0",

    # Pin to version compatible with agent_framework (0.4.14 removed LLM_REQUEST_MODEL)
    "opentelemetry-semantic-conventions-ai==0.4.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
